add logout support
