+++ src/index.js
@@ -3,7 +3,6 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 import {BrowserRouter} from "react-router-dom";
-import {ProSidebarProvider} from "react-pro-sidebar";
 import {AuthContextProvider} from "./scenes/auth/AuthContext";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
@@ -11,9 +10,7 @@ root.render(
     <React.StrictMode>
         <BrowserRouter>
             <AuthContextProvider>
-                <ProSidebarProvider>
-                    <App/>
-                </ProSidebarProvider>
+                <App/>
             </AuthContextProvider>
         </BrowserRouter>
     </React.StrictMode>
